Clean up FilterTab component
Create proper directories for FilterTabList and FilterTab components and for its
context.
